(+) Adjusted turrets' manual rotation speed to made it easier to target. (+) Added x64 build target not buildable though. (*) Corrected build time recalculation when built. (*) Renamed tasks in Sceptor. (*) Working with Sceptor serialization bug with a Docker as mother.
git-svn-id: svn://localhost/gltestplus/trunk@156 42c370b9-4569-d348-85c3-ab3aefc53057

+++ serial_util.cpp
@@ -138,7 +138,7 @@ SerializeStream *BinSerializeStream::substream(){
 }
 
 void BinSerializeStream::join(tt *o){
-	*this << ((BinSerializeStream*)o)->size;
+	*this << (unsigned long)((BinSerializeStream*)o)->size;
 	write(*(BinSerializeStream*)o);
 }
